src/main/python/requirement.txt
Flask==0.11.1
psycopg2==2.6.2
Flask-Login==0.3.2
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
flask-restful-swagger==0.19
SQLAlchemy-Searchable==0.10.2
marshmallow==2.10.0
marshmallow-sqlalchemy==0.10.0
Flask-SocketIO==2.7.1
eventlet==0.19.0
GeoAlchemy2==0.3.0
